// app/wrs-ufl/src/core/interfaces/store-user-feedback-payload.interface.ts
export interface IStoreUserFeedbackPayload {
  id: string;
  wrNumber: string;
  facility: string;
  uCode: string;
  workAgainstCode: string;
  troubleBrkdwn: string;
  opsReview: string;
  equipGroupType: string;
  equipClass: string;
  nucApplicableMode: string;
  fid: string;
  spv: string;
  mrule: string;
  ist: string;
  cda: string;
  safeSd: string;
  ePlan: string;
  fireProt: string;
  fireProt2: string;
  apprFire: string;
  flex: string;
  licRenew: string;
  equipBin: string;
  targetOutageInd: string;
  targetPriority: string;
  eCode: string;
  equipmentNumber: string;
  equipmentName: string;
  equipmentType: string;
  systemCode: string;
  locationDesc: string;
  description: string;
  detailedDescription: string;
  eCodeOutageYes: string;
  eCodeOutageNo: string;
  equipTypeOutageYes: string;
  equipTypeOutageNo: string;
  equipGroupIdOutageYes: string;
  equipGroupIdOutageNo: string;
  eCodeTrblBrkYes: string;
  eCodeTrblBrkNo: string;
  eCodeAverageTrblBrk12Mo: string;
  eCodeMaxTrblBrk12Mo: string;
  eCodeMinTrblBrk12Mo: string;
  modelName: string;
  modelVersion: string;
  modelRecommendation: string;
  modelProbability: string;
  userDecision: string;
  createUserId?: string;
  updateUserId?: string;
  slid: string;
}